UI Updates

+++ api/src/main/groovy/com/capstone/api/service/RoutingService.groovy
@@ -8,7 +8,7 @@ import org.locationtech.jts.geom.Polygon
 import org.opengis.feature.simple.SimpleFeature
 
 interface RoutingService {
-    List<RouteSegment> generateRoute(int profile_id, int src, int dst)
+    List<RouteSegment> generateRoute(int profile_id, long src, long dst)
     List<Routing_Wayzone> getWayzones(int profile_id, boolean source)
     List<InterdictionZone> generateCoverRegions(GenCoverParams params)
 }
